+++ Projet_Manic_Shooter/Assets/Scripts/Froleur.cs
@@ -9,28 +9,39 @@ public class Froleur : MonoBehaviour
  * SORTIE   : Le gain du score et l'absorption le cas échéant.
 \********************************************************/
 {
-    public GameObject o_Joueur;
+    public GameObject o_Joueur = null;
+    public CompteursJoueur c_CompteurJoueur = null;
 
     // Start is called before the first frame update
     void Start()
+    //BUT : Initialiser les références.
     {
         if (o_Joueur == null)
         {
             o_Joueur = GameObject.Find("Joueur");
         }
+
+        if (c_CompteurJoueur == null)
+        {
+            c_CompteurJoueur = GameObject.Find("GameManager").GetComponent<CompteursJoueur>();
+        }
     }
 
     void OnTriggerEnter(Collider other)
+    //BUT : Gérer la collision du frôleur avec les attaques ennemies.
     {
         if (other.gameObject.layer == LayerMask.NameToLayer("Danmaku"))
         {
             if (other.tag == o_Joueur.tag)
             {
-                Debug.Log("Absorption");
+                //Debug.Log("Absorption");
+                Destroy(other.gameObject);
+                c_CompteurJoueur.ChangeScore(2);
             }
             else
             {
                 Debug.Log("Frôlement");
+                c_CompteurJoueur.ChangeScore(1);
             }
         }
     }